Relocate views
