+++ UI/src/views/Warehouses.vue
